+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## v0.6.0 - 2021-07-
 
+- `profile`:
+    - fix a bug when no taxid mapping data given.
 
 ## v0.5.0 - 2021-06-24