+++ CHANGELOG.md
@@ -2,9 +2,12 @@
 
 ## v0.6.0 - 2021-07-
 
+- new command:
+    - `merge`: 
 - `profile`:
     - fix a nil pointer bug when no taxid mapping data given.
     - fix number of ureads.
+
 ## v0.5.0 - 2021-06-24
 
 - `compute`: